+++ .github/workflows/ci-dev.yaml
@@ -84,7 +84,7 @@ jobs:
         id: vars
         with:
           script: |
-            const { default: setupVars } = await import('${{ github.workspace }}/.github/scripts/setup-vars.js')
+            const setupVars = await import('${{ github.workspace }}/.github/scripts/setup-vars.js')
 
             const setup = JSON.parse(process.env.CI_SETUP);
             return await setupVars({core, setup})
